+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -7,25 +7,6 @@
 
 import Foundation
 
-final class FeedCachePolicy {
-    
-    private init() {}
-    
-    private static let calendar = Calendar(identifier: .gregorian)
-    
-    private static var maxCacheInAgeDays: Int {
-        return 7
-    }
-    
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxAge = calendar.date(byAdding: .day, value: maxCacheInAgeDays, to: timestamp) else {
-            return false
-        }
-        
-        return date < maxAge
-    }
-}
-
 public class LocalFeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
